requirements.txt
pandas
numpy
Pillow
torch
torchvision
matplotlib
scikit-learn
seaborn
tensorflow
torchsummary
opencv-python
mtcnn
tensorboard
